# lab6/openapi.yaml
openapi: 3.0.3
info:
  title: Comment
  description: Comment API
  version: 1.0.0
servers:
  - url: http://localhost:8086/api/v1
tags:
  - name: comment
paths:
  /comments:
    post:
      tags:
        - comment
      summary: Create new comment
      operationId: createComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Bad request
        '422':
          description: Validation error
  /comments/{commentId}:
    get:
      tags:
        - comment
      summary: Get comment by ID
      operationId: getCommentById
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          description: Comment not found
    put:
      tags:
        - comment
      summary: Update comment
      operationId: updateComment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentUpdate'
      responses:
        '200':
          description: Comment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          description: Comment not found
        '422':
          description: Validation error
    delete:
      tags:
        - comment
      summary: Delete comment
      operationId: deleteComment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Comment deleted
        '404':
          description: Comment not found

components:
  schemas:
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        content:
          type: string
          example: "This is an example comment"
        author_id:
          type: integer
          example: 123
        post_id:
          type: integer
          example: 456
        created_at:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-10-01T12:30:00Z"
      required:
        - id
        - content
        - author_id
        - post_id
        - created_at

    CommentCreate:
      type: object
      properties:
        content:
          type: string
        author_id:
          type: integer
        post_id:
          type: integer
      required:
        - content
        - author_id
        - post_id

    CommentUpdate:
      type: object
      properties:
        content:
          type: string
          enum: [active, deleted]
      required:
        - content